@@ -3,34 +3,35 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.7.1",
-    "@reduxjs/toolkit": "^1.7.2",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "@types/jest": "^24.9.1",
-    "@types/node": "^12.20.43",
-    "@types/react": "^16.14.22",
-    "@types/react-dom": "^16.9.14",
-    "@types/react-redux": "^7.1.22",
-    "@typescript-eslint/eslint-plugin": "^5.10.2",
-    "antd": "^4.18.5",
-    "axios": "^0.25.0",
-    "axios-mock-adapter": "^1.20.0",
+    "@emotion/react": "^11.9.0",
+    "@reduxjs/toolkit": "^1.8.2",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^14.2.0",
+    "@types/jest": "^28.1.1",
+    "@types/node": "^17.0.40",
+    "@types/react": "^18.0.12",
+    "@types/react-dom": "^18.0.5",
+    "@types/react-redux": "^7.1.24",
+    "@typescript-eslint/eslint-plugin": "^5.27.1",
+    "antd": "^4.21.0",
+    "axios": "^0.27.2",
+    "axios-mock-adapter": "^1.21.1",
     "classnames": "^2.3.1",
-    "eslint-config-prettier": "^8.3.0",
+    "drawflow": "^0.0.58",
+    "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.28.0",
-    "eslint-plugin-react-hooks": "^4.3.0",
-    "i18next": "^21.6.10",
-    "prettier": "^2.5.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-i18next": "^11.15.3",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "^2.1.3",
-    "typescript": "~4.1.5"
+    "eslint-plugin-react": "^7.30.0",
+    "eslint-plugin-react-hooks": "^4.5.0",
+    "i18next": "^21.8.8",
+    "prettier": "^2.6.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-i18next": "^11.17.0",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
+    "typescript": "~4.7.3"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -59,7 +60,8 @@
     ]
   },
   "devDependencies": {
-    "husky": "^7.0.4",
-    "webpack": "^5.68.0"
+    "@types/drawflow": "^0.0.4",
+    "husky": "^8.0.1",
+    "webpack": "^5.73.0"
   }
 }
