@@ -1,5 +1,5 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App'
 import { store } from './app/store'
@@ -10,11 +10,13 @@ import en from './i18n/translations/en.json'
 
 initTranslation({ en }, 'en', 'en')
 
-ReactDOM.render(
-  <React.StrictMode>
-    <Provider store={store}>
-        <App />
-    </Provider>
-  </React.StrictMode>,
-  document.getElementById('root'),
+const container = document.getElementById('root')
+const root = createRoot(container!)
+root.render(
+  // TODO: investigate how to fix https://andreasheissenberger.medium.com/react-components-render-twice-any-way-to-fix-this-91cf23961625
+  // <React.StrictMode>
+  <Provider store={store}>
+    <App />
+  </Provider>,
+  // </React.StrictMode>,
 )
