@@ -3,6 +3,6 @@ import * as types from '../types'
 
 export const getPipelineById = async (id: string) => {
   const path = buildPath(types.APIEndpoint.PipelineById, { id })
-  const { data } = await baseAPI.get<types.Pipeline[]>(path)
+  const { data } = await baseAPI.get<types.PipelineResponse>(path)
   return data
 }
