@@ -2,21 +2,18 @@ import { createAsyncThunk, createSlice, PayloadAction } from '@reduxjs/toolkit'
 import api from 'src/api'
 import { Pipeline } from 'src/api/types'
 
-export const fetchPipeline = createAsyncThunk(
-  'pipeline/fetchPipeline',
-  async (id: string, thunkAPI) => {
-    const response = await api.getPipelineById(id)
-    return response
-  }
-)
+export const fetchPipeline = createAsyncThunk('pipeline/fetchPipeline', async (id: string, thunkAPI) => {
+  const { data } = await api.getPipelineById(id)
+  return data
+})
 
 interface PipelineState {
-  items: Pipeline[]
+  data: Pipeline
   loading: boolean
 }
 
 const initialState: PipelineState = {
-  items: [],
+  data: {},
   loading: true,
 }
 
@@ -25,8 +22,8 @@ const pipelineSlice = createSlice({
   initialState,
   reducers: {},
   extraReducers: (builder) => {
-    builder.addCase(fetchPipeline.fulfilled, (state, action: PayloadAction<Pipeline[]>) => {
-      state.items = action.payload
+    builder.addCase(fetchPipeline.fulfilled, (state, action: PayloadAction<Pipeline>) => {
+      state.data = action.payload
     })
   },
 })
