@@ -1,6 +1,7 @@
 /** @jsxImportSource @emotion/react */
 
 import { Menu as AntdMenu } from 'antd'
+import { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Link, useLocation } from 'react-router-dom'
 import './Menu.css'
@@ -8,15 +9,28 @@ import './Menu.css'
 function Menu() {
   const { t } = useTranslation()
   const location = useLocation()
+
+  const menuItems = useMemo(
+    () => [
+      {
+        key: 'dashboard',
+        label: <Link to='/dashboard'>{t('routes.dashboard')}</Link>,
+      },
+      {
+        key: 'editor',
+        label: <Link to='/editor'>{t('routes.editor')}</Link>,
+      },
+    ],
+    [],
+  )
+
   return (
-    <AntdMenu selectedKeys={[location.pathname]} mode='horizontal' css={{ flex: 1, border: 'none' }}>
-      <AntdMenu.Item key='/dashboard'>
-        <Link to='/dashboard'>{t('routes.dashboard')}</Link>
-      </AntdMenu.Item>
-      <AntdMenu.Item key='/issues'>
-        <Link to='/issues'>{t('routes.issues')}</Link>
-      </AntdMenu.Item>
-    </AntdMenu>
+    <AntdMenu
+      selectedKeys={[location.pathname]}
+      mode='horizontal'
+      css={{ flex: 1, border: 'none' }}
+      items={menuItems}
+    />
   )
 }
 
